Declare debug topics in advance Potential fix for #31
Here we 'declare' each topic in the relevant module using nodebug/1
(see http://eu.swi-prolog.org/pldoc/man?predicate=debug/1)

In principle this provides some modularity, in practice, we
don't declare the md5 topic in the md5hash module. This is because
this module is loaded using ensure_loaded/1, post-compilation,
and we may want to call debug/1 on this prior to loading.
 Please enter the commit message for your changes. Lines starting